indexer/requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
gensim==4.3.2
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
idna==3.6
mecab-python3==1.0.8
numpy==1.26.3
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
scipy==1.11.4
smart-open==6.4.0
SQLAlchemy==2.0.25
sqlalchemy-cockroachdb==2.0.2
typing_extensions==4.9.0
unidic-lite==1.0.8
urllib3==2.1.0
voyager==2.0.2
